+++ Project Aela/3D/Models/Model.h
@@ -22,17 +22,17 @@ namespace Aela {
 
 			std::vector<SubModel>* getSubModels();
 			void setSubModels(std::vector<SubModel>* subModels);
-			unsigned long long getNumberOfVertices();
-			unsigned long long getNumberOfUVs();
-			unsigned long long getNumberOfNormals();
-			unsigned long long getNumberOfIndices();
+			size_t getNumberOfVertices();
+			size_t getNumberOfUVs();
+			size_t getNumberOfNormals();
+			size_t getNumberOfIndices();
 
 		private:
 			std::vector<SubModel> subModels;
 
-			unsigned long long numberOfVertices = 0;
-			unsigned long long numberOfUVs = 0;
-			unsigned long long numberOfNormals = 0;
-			unsigned long long numberOfIndices = 0;
+			size_t numberOfVertices = 0;
+			size_t numberOfUVs = 0;
+			size_t numberOfNormals = 0;
+			size_t numberOfIndices = 0;
 	};
 }
